refactor(puzzle): unify reverseBezierPath in bezier-path module (closes #279)
`reverseBezierPath` was defined in both `procedural-generator.ts` and
`composable/tab-shapes.ts`. Both implementations produced byte-identical
output on valid paths, but having two copies of a math primitive is a
latent bug source.

Move `BezierPath`, `reverseBezierPath`, and `mirrorBezierPathY` into
`composable/bezier-path.ts` (alongside `bezierPathToSvg` / `fmt`).
Update `compose.ts`, `tab-merge.ts`, `curve-clamp.ts`, and
`procedural-generator.ts` to import from the shared module, and split
the affected tests into a new `bezier-path.test.ts`.
